Reorganize scalaz compatibility module
